\name{nt_style}
\alias{nt_style}
\title{
style function for ggplot in newstapa
}
\usage{
nt_style(panel = TRUE, gridLines = "x", legend = "top")
}
\description{minimal style guides in ggplot}
\arguments{
  \item{panel}{
whether you will add background color on panel or not
}
  \item{gridLines}{
x, y, both, none ... grid you want to draw
}
  \item{legend}{
set legend posigion... top, bottom, left, right
}
}
\author{
jiyeon kim
}

